requirements.txt
telethon
openai-whisper
torch
google-generativeai
notion-client
python-dotenv